backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Future dependencies (commented for now, will be added in later tasks)
# langchain==0.1.0
# openai==1.3.0
# pandas==2.1.4
# duckdb==0.9.2
# celery==5.3.4
# redis==5.0.1
# psycopg2-binary==2.9.9
# python-multipart==0.0.6 